Touchable text and truncate refactor
