+++ lib/components/private/Truncate/Truncate.css.ts
@@ -1,0 +1,6 @@
+import { style } from '@vanilla-extract/css';
+
+export const truncate = style({
+  whiteSpace: 'nowrap',
+  textOverflow: 'ellipsis',
+});
